Added aiida-fireengine to the repostiory
This plugin provided a scheduler that does not interface to any
scheduler installed onto the remote computer, but instead interface
with the `fireworks` package to allow launching AiiDA CalcJobs to run
on the remote computer in a more flexible way. For example, multiple
AiiDA jobs can be encapsulated into a single resource allocation (of one or
more computing nodes).