chore: deprecate `development_status` key
Since the development status is now parsed from the various
configuration file formats in the plugin repository, duplicating this
information on the registry level leads to misalignment.

The information at the registry level can be dropped - for the moment,
we retain the key in cases where a corresponding trove classifier is not
yet present in the plugin repository (in order not to lose any
information).