feat: sort plugins by commit count (#241)
Use github API to get the commits endpoint.
To get the number of commits in the last year we specify the period in the params of the request
and then extract the number of pages from the response header.

For plugins not hosted on github, we clone the repository and use git directly.