Improve on proper caching for plugin metadata (#287)
The cache was still not working as expected after I removed all cached data from GH dashboard. After this change, I think it is now caching properly:

The cache is independent of PRs, for webpage build the caching is always invalid.
The condition to run metadata generating has one more condition when caching is not hit for none record in the cache database.