pr-review from forked repo (#296)
This change should make it possible to show pr-preview by PR from the forked repo, and set the `paths` parameters to `plugins.yaml` so it only can be triggered when this file changes.
The `pull_request_target` replaces the `pull_requests` event to limit the preview trigger condition to avoid malicious preview trigger actions.