Seperate the dev CI tests to run for dev changes
Fixes https://github.com/aiidateam/aiida-registry/pull/306

The CI action not triggered because the changes are made not from `plugins.yaml` file, thus not trigger the CI.
The following restriction is set to only trigger the preview CI to prevent the malicious code inject to CI actions.
https://github.com/aiidateam/aiida-registry/blob/c306d6a7cb85ef12eba51a24d4d3f21805e70b3d/.github/workflows/ci.yml#L10-L11

In this PR, I separate the actions that test changes on other development changes to other CI which will not term out to be a preview from our official aiida official domain. Besides, I remove the pre-commit action and make it triggered from GHA integration.