Add load_profile to entrypoints analyse (#314)
aiida-aimall show an error that actually can be avoid from analyse_entrypoint.py. It raises because it use default=Int(1) instead of default=lambda: Int(1) that is better in practice (see https://aiida.readthedocs.io/projects/aiida-core/en/latest/topics/processes/usage.html#validation-and-defaults). The point is we should not force developer to do this and the default=Int(1) is acceptable.

So In this PR I just add aiida.load_profile() since the profile is exist in the aiida-core-with-servecs image. In the future, we can considered to use the light profile for entry_point analysis.